backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openpyxl==3.1.5
python-multipart==0.0.20
google-api-python-client==2.159.0
google-auth==2.37.0
